+++ src/main/java/com/source/plusutil/service/noticeService/NoticeService.kt
@@ -8,7 +8,7 @@ interface NoticeService {
     fun getNoticeList(authentication: Authentication?, currentPage: Int?): NoticeListDto?
     fun writeNotice(noticeWriteDto: NoticeWriteRequestDto, authentication: Authentication?) : NoticeWriteResponseDto
     fun getNoticeTotalPage(currentPage: Int?, listSize: Int?): Int?
-    fun getNoticeDetailInfo(authentication: Authentication?, noticeNo: Int?): NoticeDetailDto?
-    fun deleteNoticeInfo(noticeNo: Int?, currentPage: Int?) : NoticeDeleteResponseDto?
-    fun updateNoticeInfo(request: HttpServletRequest?, authentication: Authentication?, noticeNo: Int?, noticeTitle: String?, noticeContent: String?, category: String?)
+    fun getNoticeDetailInfo(authentication: Authentication?, noticeNo: Long?): NoticeDetailDto?
+    fun deleteNoticeInfo(noticeNo: Long?, currentPage: Int?) : NoticeDeleteResponseDto?
+    fun updateNoticeInfo(request: HttpServletRequest?, authentication: Authentication?, noticeNo: Long?, noticeTitle: String?, noticeContent: String?, category: String?)
 }
